requirements.txt
openpyxl==3.1.2
pandas==2.1.3
pyyaml==6.0.1
rapidfuzz==3.5.2
pytest==7.4.3
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7